QSV: cast width*height to uint32_t to fix sign-compare warning (Fixes #10200)

+++ plugins/obs-qsv11/QSV_Encoder_Internal.cpp
@@ -347,7 +347,7 @@ mfxStatus QSV_Encoder_Internal::InitParams(qsv_param_t *pParams, enum qsv_codec 
 
 	constexpr uint32_t pixelcount_4k = 3840 * 2160;
 	/* If size is 4K+, set tile columns per frame to 2. */
-	if (codec == QSV_CODEC_AV1 && (pParams->nWidth * pParams->nHeight) >= pixelcount_4k) {
+	if (codec == QSV_CODEC_AV1 && static_cast<uint32_t>(pParams->nWidth * pParams->nHeight) >= pixelcount_4k) {
 		memset(&m_ExtAv1TileParam, 0, sizeof(m_ExtAv1TileParam));
 		m_ExtAv1TileParam.Header.BufferId = MFX_EXTBUFF_AV1_TILE_PARAM;
 		m_ExtAv1TileParam.Header.BufferSz = sizeof(m_ExtAv1TileParam);
